requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
dotenv==0.9.9
psycopg2-binary==2.9.10
python-dotenv==1.1.1
sqlparse==0.5.3
typing_extensions==4.14.1
